+++ docs/changelog.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## Added
+
+- Support disable show options in Select component
+
 ## [v0.4.2] - 14.03.2025
 
 ## Fixed